requirements.txt
Flask
deep-translator
requests
beautifulsoup4
gunicorn
trafilatura
newspaper3k
huggingface-hub
lxml_html_clean
